+++ temporalio/nexus/__init__.py
@@ -9,6 +9,7 @@ See https://github.com/temporalio/sdk-python/tree/main#nexus
 from ._decorators import workflow_run_operation as workflow_run_operation
 from ._operation_context import Info as Info
 from ._operation_context import LoggerAdapter as LoggerAdapter
+from ._operation_context import NexusCallback as NexusCallback
 from ._operation_context import (
     WorkflowRunOperationContext as WorkflowRunOperationContext,
 )
