+++ temporalio/nexus/_decorators.py
@@ -16,16 +16,10 @@ from nexusrpc.handler import (
     StartOperationContext,
 )
 
-from temporalio.nexus._operation_context import (
-    WorkflowRunOperationContext,
-)
-from temporalio.nexus._operation_handlers import (
-    WorkflowRunOperationHandler,
-)
-from temporalio.nexus._token import (
-    WorkflowHandle,
-)
-from temporalio.nexus._util import (
+from ._operation_context import WorkflowRunOperationContext
+from ._operation_handlers import WorkflowRunOperationHandler
+from ._token import WorkflowHandle
+from ._util import (
     get_callable_name,
     get_workflow_run_start_method_input_and_output_type_annotations,
     set_operation_factory,
