+++ temporalio/nexus/_operation_handlers.py
@@ -22,7 +22,6 @@ from nexusrpc.handler import (
     StartOperationResultAsync,
 )
 
-from temporalio import client
 from temporalio.nexus._operation_context import (
     _temporal_cancel_operation_context,
 )
@@ -67,15 +66,14 @@ class WorkflowRunOperationHandler(OperationHandler[InputT, OutputT]):
         """Start the operation, by starting a workflow and completing asynchronously."""
         handle = await self._start(ctx, input)
         if not isinstance(handle, WorkflowHandle):
-            if isinstance(handle, client.WorkflowHandle):
-                raise RuntimeError(
-                    f"Expected {handle} to be a nexus.WorkflowHandle, but got a client.WorkflowHandle. "
-                    f"You must use WorkflowRunOperationContext.start_workflow "
-                    "to start a workflow that will deliver the result of the Nexus operation, "
-                    "not client.Client.start_workflow."
-                )
             raise RuntimeError(
                 f"Expected {handle} to be a nexus.WorkflowHandle, but got {type(handle)}. "
+                f"When using @workflow_run_operation you must use "
+                "WorkflowRunOperationContext.start_workflow() "
+                "to start a workflow that will deliver the result of the Nexus operation, "
+                "and you must return the nexus.WorkflowHandle that it returns. "
+                "It is not possible to use client.Client.start_workflow() and client.WorkflowHandle "
+                "for this purpose."
             )
         return StartOperationResultAsync(handle.to_token())
